+++ template-alpine-js/src/main.js
@@ -1,0 +1,10 @@
+import Alpine from 'alpinejs'
+import './main.css'
+
+window.Alpine = Alpine
+
+Alpine.store('shop', {
+  crx: 'create-chrome-ext',
+})
+
+Alpine.start()